Separate out database interaction logic
