+++ MonitoringService/Services/Logging.cs
@@ -21,6 +21,11 @@ namespace MonitoringService.Services
             _logger.LogInformation(message);
         }
 
+        public void LogWarning(string message)
+        {
+            _logger.LogWarning(message);
+        }
+
         public void LogError(string message)
         {
             _logger.LogError(message);
