+++ MonitoringService/Worker.cs
@@ -19,6 +19,7 @@ namespace MonitoringService
         private readonly NewFileManagment _newFileManagment;
         private readonly ParsePdfs _parsePdfs;
         private readonly SpecDetailsManagement _specDetailsManagement;
+        private readonly SpecDbOperations _specDbOperations;
         private readonly string _ongoingFolderPath;
         private readonly string _approvedFolderPath;
         private readonly string _approvedCsvPath;
@@ -32,7 +33,7 @@ namespace MonitoringService
         private readonly List<string> _previousOngoingFiles;
         private readonly List<string> _previousApprovedFiles;
 
-        public Worker(ILogger<Worker> logger, IOptions<ConfigurableSettings> folderSettings, FileDirectorySetup fileDirectorySetup , NewFileManagment newFileManagment, ParsePdfs parsePdfs, SpecDetailsManagement specDetailsManagement, EmailService emailService, IServiceScopeFactory serviceScopeFactory)
+        public Worker(ILogger<Worker> logger, IOptions<ConfigurableSettings> folderSettings, FileDirectorySetup fileDirectorySetup , NewFileManagment newFileManagment, ParsePdfs parsePdfs, SpecDetailsManagement specDetailsManagement, SpecDbOperations specDbOperations, EmailService emailService, IServiceScopeFactory serviceScopeFactory)
         {
             _logger = logger;
             _ongoingFolderPath = folderSettings.Value.Ongoing;
@@ -44,12 +45,13 @@ namespace MonitoringService
             _newFileManagment = newFileManagment;
             _parsePdfs = parsePdfs;
             _specDetailsManagement = specDetailsManagement;
+            _specDbOperations = specDbOperations;
 
             _previousOngoingFileNamesPath = "previousOngoingFiles.txt";
             _previousApprovedFileNamesPath = "previousApprovedFiles.txt";
             //_previousOngoingFiles = LoadPreviousFileNames(_previousOngoingFileNamesPath);
-            _previousOngoingFiles = GetFileNamesFromDatabase("Ongoing");
-            _previousApprovedFiles = GetFileNamesFromDatabase("Approved");
+            _previousOngoingFiles = _specDbOperations.GetFileNamesFromDatabase("Ongoing");
+            _previousApprovedFiles = _specDbOperations.GetFileNamesFromDatabase("Approved");
             //_previousApprovedFiles = LoadPreviousFileNames(_previousApprovedFileNamesPath);
             _emailService = emailService;
             //_dbContext = context;
@@ -72,40 +74,7 @@ namespace MonitoringService
             File.WriteAllLines(filePath, files);
         }
 
-        private List<string> GetFileNamesFromDatabase(string Folder)
-        {
-            using var scope = _serviceScopeFactory.CreateScope();
-            var dbContext = scope.ServiceProvider.GetRequiredService<ApplicationContext>();
-
-            return dbContext.SpecDetails.Where(s => s.Folder == Folder).Select(s => s.FileName).ToList();
-        }
-
-        public List<SpecDetails> GetDatabaseEntries(string folder)
-        {
-            using var scope = _serviceScopeFactory.CreateScope();
-            var dbContext = scope.ServiceProvider.GetRequiredService<ApplicationContext>();
-            List<SpecDetails> entries = new List<SpecDetails>();
-
-            if (folder == "Ongoing")
-            {
-                entries = dbContext.SpecDetails.Where(s => s.Folder == "Ongoing").ToList();
-            }
-            else if (folder == "Approved")
-            {
-                entries = dbContext.SpecDetails.Where(s => s.Folder == "Approved").ToList();
-            }
-
-            List<SpecDetails> existingSpecs = new List<SpecDetails>();
-
-            Console.WriteLine("Database Entries:");
-            foreach (var entry in entries)
-            {
-                Console.WriteLine($"ID: {entry.Id}, Title: {entry.Title}, Author: {entry.Author}");
-                existingSpecs.Add(entry);
-            }
-
-            return existingSpecs;
-        }
+       
 
 
         protected override async Task ExecuteAsync(CancellationToken stoppingToken)
@@ -162,7 +131,7 @@ namespace MonitoringService
                         if (listOngoingFiles.Count > 0)
                         {
                             _emailService.SendNewFilesEmail(listOngoingFiles, newOngoingFiles, "Ongoing");
-                            SaveToDatabase(listOngoingFiles);
+                            _specDbOperations.SaveToDatabase(listOngoingFiles);
                         }
 
                         if (emptyListOngoingFiles.Count > 0)
@@ -206,7 +175,7 @@ namespace MonitoringService
                         if (listApprovedFiles.Count > 0)
                         {
                             _emailService.SendNewFilesEmail(listApprovedFiles, newApprovedFiles, "Ongoing");
-                            SaveToDatabase(listApprovedFiles);
+                            _specDbOperations.SaveToDatabase(listApprovedFiles);
                             CreateAndSaveCsvFile(listApprovedFiles);
                         }
 
@@ -257,28 +226,7 @@ namespace MonitoringService
         //}
 
 
-        private void SaveToDatabase(List<SpecDetails> details)
-        {
-            using var scope = _serviceScopeFactory.CreateScope();
-            var dbContext = scope.ServiceProvider.GetRequiredService<ApplicationContext>();
-
-            foreach (var specRowDocument in details)
-            {
-                if (_specDetailsManagement.AllSpecFieldsEntered(specRowDocument))
-                {
-                    dbContext.SpecDetails.Add(specRowDocument);
-                }
-                else
-                {
-                    _logger.LogWarning($"Skipping document {specRowDocument.FileName} due to missing information.");
-                    var issue = $"There was an issue retrieving some information from spec {specRowDocument.FileName} in the {specRowDocument.Folder} folder. Please review to ensure spec is formatted correctly and fully complete and update the file.\n";
-                    _emailService.SendAdminErrorMail(specRowDocument.FileName, issue, specRowDocument.Folder);
-                }
-            }
-
-            dbContext.SaveChanges();
-        }
-
+
